# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NONONICK Universal AI Editor & Agent Engine API",
    "description": "Production-grade RESTful API connecting external AI models, Chatbots, Telegram/Discord bots, and developer automations with the NONONICK Universal Web Studio.",
    "version": "2.5.0",
    "contact": {
      "name": "NONONICK AI Studio Core Team",
      "email": "alirezassezar.1@gmail.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "https://ais-pre-5ilkxtfwf6yibz3jj2ynl3-17774190120.europe-west3.run.app",
      "description": "Production Cloud Gateway"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/chat/completions": {
      "post": {
        "tags": ["Chatbot & LLM Integration"],
        "summary": "OpenAI-compatible Chat Completions",
        "description": "Drop-in standard endpoint compatible with OpenAI SDKs, LangChain, OpenWebUI, and LibreChat.",
        "operationId": "createChatCompletion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": { "type": "string", "default": "gemini-3.8-flash" },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": { "type": "string", "enum": ["system", "user", "assistant"] },
                        "content": { "type": "string" }
                      },
                      "required": ["role", "content"]
                    }
                  },
                  "temperature": { "type": "number", "default": 0.7 }
                },
                "required": ["messages"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OpenAI chat completion response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "400": { "description": "Invalid request payload" }
        }
      }
    },
...
    "/api/mock/{projectId}/{collection}": {
      "get": {
        "tags": ["Mock REST API"],
        "summary": "Query records in a project's database collection",
        "operationId": "queryMockCollection",
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "collection", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "search", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "sort", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "order", "in": "query", "required": false, "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": { "description": "Array of matching collection records" },
          "404": { "description": "Project or collection not found" }
        }
      },
      "post": {
        "tags": ["Mock REST API"],
        "summary": "Create record in project's database collection",
        "operationId": "createMockRecord",
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "collection", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "201": { "description": "Record created" },
          "404": { "description": "Project or collection not found" }
        }
      }
    },
...
    "/api/auth/verify": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Verify Firebase Authentication ID Token or Session",
        "operationId": "verifyAuthToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": { "type": "string" }
                },
                "required": ["idToken"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token is valid and decoded" },
          "401": { "description": "Invalid or expired token" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
...
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string", "example": "chat.completion" },
          "created": { "type": "integer" },
          "model": { "type": "string" },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": { "type": "integer" },
                "message": {
                  "type": "object",
                  "properties": {
                    "role": { "type": "string" },
                    "content": { "type": "string" }
                  }
                },
                "finish_reason": { "type": "string" }
              }
            }
          },
          "usage": {
            "type": "object",
            "properties": {
              "prompt_tokens": { "type": "integer" },
              "completion_tokens": { "type": "integer" },
              "total_tokens": { "type": "integer" }
            }
          }
        }
      }
    }
  }
}